+++ lib/nima/actor_bone.dart
@@ -1,9 +1,9 @@
-import "readers/stream_reader.dart";
 import "actor.dart";
 import "actor_bone_base.dart";
 import "actor_component.dart";
 import "actor_node.dart";
 import "jelly_component.dart";
+import "readers/stream_reader.dart";
 
 class ActorBone extends ActorBoneBase {
   ActorBone _firstBone;
@@ -13,18 +13,20 @@ class ActorBone extends ActorBoneBase {
     return _firstBone;
   }
 
+  @override
   ActorComponent makeInstance(Actor resetActor) {
     ActorBone instanceNode = ActorBone();
     instanceNode.copyBoneBase(this, resetActor);
     return instanceNode;
   }
 
+  @override
   void completeResolve() {
     super.completeResolve();
     if (children == null) {
       return;
     }
-    for (ActorNode node in children) {
+    for (final ActorNode node in children) {
       if (node is ActorBone) {
         _firstBone = node;
         return;
@@ -33,9 +35,7 @@ class ActorBone extends ActorBoneBase {
   }
 
   static ActorBone read(Actor actor, StreamReader reader, ActorBone node) {
-    if (node == null) {
-      node = ActorBone();
-    }
+    node ??= ActorBone();
     ActorBoneBase.read(actor, reader, node);
     return node;
   }