+++ lib/nima/actor_component.dart
@@ -1,6 +1,6 @@
-import "readers/stream_reader.dart";
 import "actor.dart";
 import "actor_node.dart";
+import "readers/stream_reader.dart";
 
 abstract class ActorComponent {
   String _name = "Unnamed";
@@ -12,11 +12,8 @@ abstract class ActorComponent {
   int dirtMask = 0;
   List<ActorComponent> dependents;
 
-  ActorComponent.withActor(Actor actor) {
-    this.actor = actor;
-  }
-
   ActorComponent();
+  ActorComponent.withActor(this.actor);
 
   String get name {
     return _name;