+++ lib/nima/actor_distance_constraint.dart
@@ -1,9 +1,9 @@
 import "actor.dart";
 import "actor_node.dart";
 import "actor_targeted_constraint.dart";
-import "readers/stream_reader.dart";
-import "math/vec2d.dart";
 import "math/mat2d.dart";
+import "math/vec2d.dart";
+import "readers/stream_reader.dart";
 
 class DistanceMode {
   static const int Closer = 0;
@@ -19,9 +19,7 @@ class ActorDistanceConstraint extends ActorTargetedConstraint {
 
   static ActorDistanceConstraint read(
       Actor actor, StreamReader reader, ActorDistanceConstraint component) {
-    if (component == null) {
-      component = ActorDistanceConstraint();
-    }
+    component ??= ActorDistanceConstraint();
     ActorTargetedConstraint.read(actor, reader, component);
 
     component._distance = reader.readFloat32("distance");
@@ -44,13 +42,13 @@ class ActorDistanceConstraint extends ActorTargetedConstraint {
   }
 
   @override
-  constrain(ActorNode node) {
-    ActorNode t = this.target;
+  void constrain(ActorNode node) {
+    ActorNode t = target as ActorNode;
     if (t == null) {
       return;
     }
 
-    ActorNode p = this.parent;
+    ActorNode p = parent;
     Vec2D targetTranslation = t.getWorldTranslation(Vec2D());
     Vec2D ourTranslation = p.getWorldTranslation(Vec2D());
 
@@ -84,23 +82,25 @@ class ActorDistanceConstraint extends ActorTargetedConstraint {
     world[5] = position[1];
   }
 
+  @override
   void update(int dirt) {}
+  @override
   void completeResolve() {}
 
-  get distance => _distance;
-  get mode => _mode;
+  double get distance => _distance;
+  int get mode => _mode;
 
   set distance(double d) {
     if (_distance != d) {
       _distance = d;
-      this.markDirty();
+      markDirty();
     }
   }
 
   set mode(int m) {
     if (_mode != m) {
       _mode = m;
-      this.markDirty();
+      markDirty();
     }
   }
 }