+++ lib/nima/math/aabb.dart
@@ -1,5 +1,5 @@
-import "dart:typed_data";
 import "dart:math";
+import "dart:typed_data";
 import "vec2d.dart";
 
 class AABB {
@@ -10,11 +10,11 @@ class AABB {
   }
 
   AABB() {
-    this._buffer = Float32List.fromList([0.0, 0.0, 0.0, 0.0]);
+    _buffer = Float32List.fromList([0.0, 0.0, 0.0, 0.0]);
   }
 
   AABB.clone(AABB a) {
-    this._buffer = Float32List.fromList(a.values);
+    _buffer = Float32List.fromList(a.values);
   }
 
   AABB.fromValues(double a, double b, double c, double d) {
@@ -22,11 +22,11 @@ class AABB {
   }
 
   double operator [](int idx) {
-    return this._buffer[idx];
+    return _buffer[idx];
   }
 
-  operator []=(int idx, double v) {
-    this._buffer[idx] = v;
+  void operator []=(int idx, double v) {
+    _buffer[idx] = v;
   }
 
   static AABB copy(AABB out, AABB a) {