+++ lib/nima/math/vec2d.dart
@@ -1,5 +1,5 @@
-import "dart:typed_data";
 import "dart:math";
+import "dart:typed_data";
 import "mat2d.dart";
 
 class Vec2D {
@@ -96,8 +96,8 @@ class Vec2D {
   }
 
   static Vec2D negate(Vec2D result, Vec2D a) {
-    result[0] = -a[0];
-    result[1] = -a[1];
+    result[0] = -1 * a[0];
+    result[1] = -1 * a[1];
 
     return result;
   }