+++ lib/nima/readers/block_reader.dart
@@ -1,19 +1,19 @@
+import "dart:typed_data";
 import "binary_reader.dart";
-import 'dart:typed_data';
 
 class BlockReader extends BinaryReader {
   @override
   int blockType;
 
   BlockReader(ByteData data) : super(data) {
-    this.blockType = 0;
+    blockType = 0;
   }
 
-  BlockReader.fromBlock(int type, ByteData stream) : super(stream) {
-    this.blockType = type;
-  }
+  BlockReader.fromBlock(this.blockType, ByteData stream) : super(stream);
 
-  // A binary block is defined as a TLV with type of one byte, length of 4 bytes, and then the value following.
+  // A binary block is defined as a TLV with type of one byte, length of 4
+  // bytes, and then the value following.
+  @override
   BlockReader readNextBlock(Map<String, int> types) {
     if (isEOF()) {
       return null;