+++ lib/nima/readers/stream_reader.dart
@@ -7,15 +7,15 @@ abstract class StreamReader {
   int blockType = 0;
 
   // Instantiate the right type of Reader based on the input values
-  factory StreamReader(data) {
-    var reader;
+  factory StreamReader(dynamic data) {
+    StreamReader reader;
     if (data is ByteData) {
       reader = BlockReader(data);
       // Move the readIndex forward for the binary reader.
-      reader.readUint8();
-      reader.readUint8();
-      reader.readUint8();
-      reader.readUint8();
+      reader.readUint8("N");
+      reader.readUint8("I");
+      reader.readUint8("M");
+      reader.readUint8("A");
     } else if (data is Map) {
       reader = JSONBlockReader(data);
     }
@@ -29,17 +29,18 @@ abstract class StreamReader {
   int readUint32Length();
 
   int readUint8(String label);
-  readUint8Array(Uint8List list, int length, int offset, String label);
+  void readUint8Array(Uint8List list, int length, int offset, String label);
   int readInt8(String label);
   int readUint16(String label);
-  readUint16Array(Uint16List ar, int length, int offset, String label);
+  void readUint16Array(Uint16List ar, int length, int offset, String label);
   int readInt16(String label);
   int readInt32(String label);
   int readUint32(String label);
   int readVersion();
   double readFloat32(String label);
-  readFloat32Array(Float32List ar, String label);
-  readFloat32ArrayOffset(Float32List ar, int length, int offset, String label);
+  void readFloat32Array(Float32List ar, String label);
+  void readFloat32ArrayOffset(
+      Float32List ar, int length, int offset, String label);
   double readFloat64(String label);
 
   String readString(String label);
@@ -50,10 +51,10 @@ abstract class StreamReader {
 
   StreamReader readNextBlock(Map<String, int> types);
 
-  openArray(String label);
-  closeArray();
-  openObject(String label);
-  closeObject();
+  void openArray(String label);
+  void closeArray();
+  void openObject(String label);
+  void closeObject();
 
   String get containerType;
 }