+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [1.0.4] - 2019-07-11 10:00:50
+
+* Fix breaking index buffer format change (from Int32List to Uint16List).
+* Lots of static analysis fixes.
+
 ## [1.0.3] - 2019-04-09 09:50:41
 
 * Making sure the Nima Actor widget is disposed of properly when the leaf render widet is unmounted or the render box is detached.